Renaming to "addon-resizer".
