+++ addon-resizer/.gitignore
