+++ addon-resizer/Dockerfile
