+++ addon-resizer/nanny/nanny_lib.go
