+++ addon-resizer/nanny/nanny_lib_test.go
